Moved DoneToken to codec package
